refactor(bridge): dissolve CreateUtility into the wrappers that front it
CreateUtility had grown to 984 lines behind four files that did nothing but
forward to it -- Scenes, Prefabs, Assets, and Project were 35, 34, 35, and 15
lines of pass-through. The seam was already drawn; this moves the bodies across
it so each command lives in the file named after its surface.

What stays in CreateUtility is only what more than one of the four uses: the
path helpers, ResolveCreateParent (Scenes and Prefabs), and EscapeJsonString
(Assets and Project). Everything with a single caller moved next to it --
ResolveScriptType, GenerateGuid, and the guid regex to Assets, FindMatchingBrace
to Project, the empty-scene template guid and AddDefaultSceneObjects to Scenes.

Pure relocation: no signature, payload, or behaviour changed. 341/194/189/179/125
lines, all under the 640 guideline.
